+ bootstrap * some fixes

+++ views/fileList.tpl
@@ -1,12 +1,24 @@
-<a href="/upload">Download file</a>
-<table>
-    {{range $key, $val := .Files}}
-        <tr>
+<table class="table table-bordered">
+    {{ if .Files}}
+        <thead>
+            <tr>
+                <td>File name</td>
+                <td>Upload time</td>
+                <td></td>
+                <td></td>
+            </tr>
+        </thead>
+        <tbody>
+        {{range $key, $val := .Files}}
+            <tr>
 
-            <td>{{$val.UserFileName}}</td>
-            <td>{{$val.UploadTime}}</td>
-            <td><a href="/download/{{$val.UploadTime}}">Download</a></td>
-            <td><a href="/delete/{{$val.UploadTime}}">Remove</a></td>
-        </tr>
+                <td>{{$val.UserFileName}}</td>
+                <td>{{$val.UploadTime}}</td>
+                <td><a href="/download/{{$val.UploadTime}}">Download</a></td>
+                <td><a href="/delete/{{$val.UploadTime}}">Remove</a></td>
+            </tr>
+        {{end}}
+        </tbody>
     {{end}}
+
 </table>
